requirements.txt
Django==1.11.7
psycopg2
djangorestframework
numpy
django-nose
coverage